update fetch online users and remove user from a room

+++ src/main/java/ch/uzh/ifi/hase/soprafs23/game/Room.java
@@ -19,7 +19,7 @@ public class Room {
     public Room(int roomId) { this.roomId = roomId; }
 
     public void addUser(long userId){ if (this.userIds.size() < 2) this.userIds.add(userId); }
-    public void removeUser(int userId){ this.userIds.remove(userIds.indexOf(userId)); }
+    public void removeUser(long userId){ this.userIds.remove(userIds.indexOf(userId)); }
 
     public void enterGame(){
         if (this.userIds.size() != 2) throw new IllegalStateException("Not enough players in the room!");
